About Put created

+++ app/crud/crud.py
@@ -16,9 +16,40 @@ class Crud:
             
         return res
     
+    def put(self,db,table,id,row):
+        res_query = db.query(table).filter(id==id)
+        
+        res = res_query.first()
+        
+        if res is None:
+            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,detail="No Data Found")
+        
+        if res.id != id:
+            raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                            detail="Not authorized to perform requested action")
+        try:
+            for key in row.dict():
+                if key != 'id':
+                    res.key = row.dict()[key]
+        except:
+            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail="Column is not here")
+            
+        res_query.update(res, synchronize_session=False)
+        db.commit()
+        return res_query.first()
+        
+
     def delete(self,db,table,id):
-        res = db.query(table).filter(id==id).all()
+        res = db.query(table).filter(id==id).first()
+        
+        if res is None:
+            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,detail="No Data Found")
         
         db.delete(res)
         db.commit()
         
+        return {"message":"Content deleted"}
+        
+
+crud_obj = Crud()
